refactor: extract controls from Panel
