+++ src/components/Panel/index.js
@@ -1,46 +1,12 @@
 import React from 'react'
 import Editor from '../Editor'
 import Result from '../Result'
-import storage, { mkStore } from '../../utils/storage'
+import Controls from '../Controls'
+import storage, { snippets } from '../../utils/storage'
 import './style.css'
 
 const wrapInTry = code => {
   return `try{ ${code} } catch (e) {e}`
-}
-
-const snippets = mkStore('snippet__')
-
-class SnippetSelector extends React.Component {
-  constructor () {
-    super()
-    this.onChange = this.onChange.bind(this)
-  }
-
-  onChange (ev) {
-    this.props.onChange(ev.target.value)
-  }
-
-  render () {
-    const snipps = snippets.ls()
-    return (
-      snipps.length > 0 ? <select onChange={this.onChange}>
-        <option>-</option>
-      {
-        snipps.map(x => <option key={x} value={x}>{ x }</option>)
-      }</select> : null
-    )
-  }
-}
-
-class Controls extends React.Component {
-  render () {
-    return (
-      <div>
-        <button onClick={this.props.onSaveSnippet}>Save snippet</button>&nbsp;
-        Load snippet : <SnippetSelector onChange={this.props.onLoadSnippet}/>
-      </div>
-    )
-  }
 }
 
 const DEFAULT_CODE = `scrape($, {
